test: rename integration .mts files to .ts (review: christian-bromann)
- git mv all integration *.mts -> *.ts (graphs + test suite + utils)
- update ./utils.mts imports, configure-langgraph.py graph paths
- add "type": "module" to graphs/package.json so the .ts graphs stay ESM
- fix lint surfaced by the stricter .ts ruleset (unused bindings, unsafe
  optional chaining), keeping side-effecting calls intact
